add alternate project types
In starting a Hokusai project for Node, I was initially unsure how to apply the directions. I searched in the repo for `ruby-rails` to see if it was obvious what other keywords could be substituted. It was, indeed, obvious, but it seems potentially more obvious to list the project types. The tradeoff would be that this list has to manually be kept in sync. Maybe a more robust solution would be to make the project types list dynamic and inspectable via the CLI? 

Anyway, just thought I'd open a PR in case this seemed like a good idea.