Update hokusai check command to find executables via which
For the `hokusai check` command, we want to verify that required
executables are available on the system. We currently call an
executable's version subcommand, but the invocation varies and it's
possible that an implementation may exit nonzero even if the program is
available.

Instead of relying on the behavior of various `version` subcommands,
simply check that the executable exists in the system `PATH`.